fixed multiple bugs
Checked with sphere n=10, I think its good now. But code needs to be improved to be more flexiple

+++ FitnessLandscapeAnalysis/Metrics.cs
@@ -89,7 +89,7 @@ namespace FitnessLandscapeAnalysis
             {
                 sumfd += (y[j] - ymean) * (d[j] - dmean);
             }
-            fdc = (sumfd / P) / (sigma_f * sigma_d);
+            fdc = (sumfd / (double)P) / (sigma_f * sigma_d);
             return fdc;
         }
